+++ application/src/main/java/org/cris6h16/facades/ProductDTO.java
@@ -20,7 +20,7 @@ public class ProductDTO {
     private BigDecimal price;
     private Integer stock;
     private String description;
-    private Integer weightPounds;
+    private BigDecimal weightPounds;
     private Integer widthCM;
     private Integer heightCM;
     private Integer lengthCM;
