+++ product/src/main/java/org/cris6h16/product/ProductEntity.java
@@ -66,7 +66,7 @@ public class ProductEntity {
     private String description;
 
     @Column(nullable = false, name = "weight_pounds")
-    private Integer weightPounds;
+    private BigDecimal weightPounds;
 
     @Column(nullable = false, name = "width_cm")
     private Integer widthCM;
