+++ product/src/main/java/org/cris6h16/product/ProductOutput.java
@@ -21,7 +21,7 @@ public class ProductOutput {
     private BigDecimal price;
     private Integer stock;
     private String description;
-    private Integer weightPounds;
+    private BigDecimal weightPounds;
     private Integer widthCM;
     private Integer heightCM;
     private Integer lengthCM;
